+++ bindgen-tests/tests/headers/default_visibility_private_respects_cxx_access_spec.h
@@ -1,0 +1,12 @@
+// bindgen-flags: --respect-cxx-access-specs --no-layout-tests --default-visibility=private 
+
+struct Point {
+    int x;
+    int y;
+};
+
+struct Color {
+    char r :1;
+    char g :1;
+    char b :1;
+};