+++ bindgen/lib.rs
@@ -61,6 +61,8 @@ mod regex_set;
 use codegen::CodegenError;
 use ir::comment;
 
+pub use ir::annotations::FieldVisibilityKind;
+
 pub use crate::codegen::{
     AliasVariation, EnumVariation, MacroTypeVariation, NonCopyUnionStyle,
 };
@@ -651,6 +653,11 @@ impl Builder {
         if let Some(ref suffix) = self.options.wrap_static_fns_suffix {
             output_vector.push("--wrap-static-fns-suffix".into());
             output_vector.push(suffix.clone());
+        }
+
+        if self.options.default_visibility != FieldVisibilityKind::Public {
+            output_vector.push("--default-visibility".into());
+            output_vector.push(self.options.default_visibility.to_string());
         }
 
         if cfg!(feature = "experimental") {
@@ -1813,6 +1820,18 @@ impl Builder {
         self.options.wrap_static_fns_suffix = Some(suffix.as_ref().to_owned());
         self
     }
+
+    /// Set the default visibility of fields, including bitfields and accessor methods for
+    /// bitfields.
+    ///
+    /// This option is ignored if the [`Builder::respect_cxx_access_specs`] method is enabled.
+    pub fn default_visibility(
+        mut self,
+        visibility: FieldVisibilityKind,
+    ) -> Self {
+        self.options.default_visibility = visibility;
+        self
+    }
 }
 
 /// Configuration options for generated bindings.
@@ -2159,6 +2178,9 @@ struct BindgenOptions {
     wrap_static_fns_suffix: Option<String>,
 
     wrap_static_fns_path: Option<PathBuf>,
+
+    /// Default visibility of structs and their fields.
+    default_visibility: FieldVisibilityKind,
 }
 
 impl BindgenOptions {
@@ -2354,6 +2376,7 @@ impl Default for BindgenOptions {
             wrap_static_fns,
             wrap_static_fns_suffix,
             wrap_static_fns_path,
+            default_visibility,
         }
     }
 }